move block contents to api crate, rename blob sidecars list
