+++ consensus/types/src/beacon_block_and_blob_sidecars.rs
@@ -1,5 +1,5 @@
 use crate::{
-    AbstractExecPayload, BeaconBlock, BlobSidecars, EthSpec, ForkName, ForkVersionDeserialize,
+    AbstractExecPayload, BeaconBlock, BlobSidecarList, EthSpec, ForkName, ForkVersionDeserialize,
 };
 use derivative::Derivative;
 use serde_derive::{Deserialize, Serialize};
@@ -11,7 +11,7 @@ use tree_hash_derive::TreeHash;
 #[serde(bound = "T: EthSpec, Payload: AbstractExecPayload<T>")]
 pub struct BeaconBlockAndBlobSidecars<T: EthSpec, Payload: AbstractExecPayload<T>> {
     pub block: BeaconBlock<T, Payload>,
-    pub blob_sidecars: BlobSidecars<T>,
+    pub blob_sidecars: BlobSidecarList<T>,
 }
 
 impl<T: EthSpec, Payload: AbstractExecPayload<T>> ForkVersionDeserialize
@@ -25,7 +25,7 @@ impl<T: EthSpec, Payload: AbstractExecPayload<T>> ForkVersionDeserialize
         #[serde(bound = "T: EthSpec")]
         struct Helper<T: EthSpec> {
             block: serde_json::Value,
-            blob_sidecars: BlobSidecars<T>,
+            blob_sidecars: BlobSidecarList<T>,
         }
         let helper: Helper<T> = serde_json::from_value(value).map_err(serde::de::Error::custom)?;
