+++ consensus/types/src/lib.rs
@@ -102,7 +102,6 @@ pub mod sqlite;
 pub mod beacon_block_and_blob_sidecars;
 pub mod blob_sidecar;
 pub mod blobs_sidecar;
-pub mod block_contents;
 pub mod signed_blob;
 pub mod signed_block_and_blobs;
 pub mod transaction;
@@ -126,9 +125,8 @@ pub use crate::beacon_block_body::{
 pub use crate::beacon_block_header::BeaconBlockHeader;
 pub use crate::beacon_committee::{BeaconCommittee, OwnedBeaconCommittee};
 pub use crate::beacon_state::{BeaconTreeHashCache, Error as BeaconStateError, *};
-pub use crate::blob_sidecar::{BlobSidecar, BlobSidecars};
+pub use crate::blob_sidecar::{BlobSidecar, BlobSidecarList};
 pub use crate::blobs_sidecar::{Blobs, BlobsSidecar};
-pub use crate::block_contents::BlockContents;
 pub use crate::bls_to_execution_change::BlsToExecutionChange;
 pub use crate::chain_spec::{ChainSpec, Config, Domain};
 pub use crate::checkpoint::Checkpoint;
